Rename setSpace to ensureSpaceIsSet
