+++ lib/commands/__specs__/ensure-space-set.spec.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const api = require('../../api')
-const { execute: subject } = require('../set-space')
+const { execute: subject } = require('../ensure-space-set')
 require('../../../test/init-api')
 const { generateSpace } = require('../../../test/mocks')
 const { sandbox } = require('../../../test/sandbox')
