+++ lib/commands/ensure-space-set.js
